# docs/swagger.yaml
definitions:
  errorHandler.HttpErr:
    properties:
      message:
        example: Error message
        type: string
      status_code:
        example: 400
        type: integer
    type: object
  input.AppealUpdate:
    properties:
      address:
        example: Улица Пушкина
        type: string
      appeal_type_id:
        example: 1
        type: integer
      description:
        example: Текст обращения
        type: string
      latitude:
        example: 54.1234
        type: number
      longitude:
        example: 122.7656
        type: number
      title:
        example: Обращение
        type: string
    type: object
...
  response.Appeal:
    properties:
      address:
        example: Улица Пушкина
        type: string
      appeal_photos:
        items:
          $ref: '#/definitions/response.AppealPhoto'
        type: array
      appeal_status:
        $ref: '#/definitions/response.AppealStatus'
      appeal_type:
        $ref: '#/definitions/response.AppealType'
      description:
        example: Текст обращения
        type: string
      id:
        example: 1
        type: integer
      latitude:
        example: 54.1234
        type: number
      longitude:
        example: 122.7656
        type: number
      title:
        example: Обращение
        type: string
      user:
        $ref: '#/definitions/response.User'
    type: object
  response.AppealCategory:
    properties:
      id:
        example: 1
        type: integer
      title:
        example: Категория
        type: string
    type: object
  response.AppealComment:
    properties:
      appeal_comment_photos:
        items:
          $ref: '#/definitions/response.AppealCommentPhoto'
        type: array
      date:
        example: "2024-02-10T00:00:00+05:00"
        type: string
      id:
        example: 1
        type: integer
      text:
        example: Текст комментария
        type: string
      user:
        $ref: '#/definitions/response.User'
    type: object
  response.AppealCommentPaged:
    properties:
      items:
        items:
          $ref: '#/definitions/response.AppealComment'
        type: array
      next_page_url:
        example: /api/news?page=1&page_size=10
        type: string
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
    type: object
  response.AppealCommentPhoto:
    properties:
      id:
        example: 1
        type: integer
      url:
        example: https://storage.yandexcloud.net/urbathon/test.jpg
        type: string
    type: object
  response.AppealPaged:
    properties:
      items:
        items:
          $ref: '#/definitions/response.Appeal'
        type: array
      next_page_url:
        example: /api/news?page=1&page_size=10
        type: string
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
    type: object
  response.AppealPhoto:
    properties:
      id:
        example: 1
        type: integer
      url:
        example: https://storage.yandexcloud.net/urbathon/test.jpg
        type: string
    type: object
  response.AppealStatus:
    properties:
      id:
        example: 1
        type: integer
      status:
        example: Решено
        type: string
    type: object
  response.AppealType:
    properties:
      appeal_category:
        $ref: '#/definitions/response.AppealCategory'
      id:
        example: 1
        type: integer
      title:
        example: Тип
        type: string
    type: object
  response.AppealTypeByCategory:
    properties:
      id:
        example: 1
        type: integer
      title:
        example: Тип
        type: string
    type: object
...
  response.Organization:
    properties:
      address:
        type: string
      id:
        example: 1
        type: integer
      inn:
        type: string
      name:
        example: ЖКХ
        type: string
      phone:
        type: string
    type: object
...
  response.User:
    properties:
      email:
        example: test@gmail.com
        type: string
      first_name:
        example: Иван
        type: string
      id:
        example: 1
        type: integer
      job:
        type: string
      last_name:
        example: Иванов
        type: string
      organization:
        $ref: '#/definitions/response.Organization'
      organization_id:
        example: 1
        type: integer
      patronymic:
        example: Иванович
        type: string
    type: object
info:
  contact: {}
  description: Спецификация приложения команды подCRUDули
  title: Urbathon-2023
paths:
  /appeal:
    get:
      description: получение всех обращений
      parameters:
      - example: name
        in: query
        name: field
        type: string
      - in: query
        name: lat_down
        type: number
      - in: query
        name: lat_up
        type: number
      - in: query
        name: long_down
        type: number
      - in: query
        name: long_up
        type: number
      - example: asc
        in: query
        name: order
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      - in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AppealPaged'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: получение всех обращений
      tags:
      - appeal
    post:
      consumes:
      - multipart/form-data
      description: Создание обращения
      parameters:
      - example: Улица Пушкина
        in: formData
        name: address
        type: string
      - example: 1
        in: formData
        name: appeal_type_id
        type: integer
      - example: Текст обращения
        in: formData
        name: description
        type: string
      - example: 54.1234
        in: formData
        name: latitude
        type: number
      - example: 122.7656
        in: formData
        name: longitude
        type: number
      - example: Обращение
        in: formData
        name: title
        type: string
      - collectionFormat: csv
        description: photos
        in: formData
        items:
          type: file
        name: photos
        required: true
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Appeal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Создание обращения
      tags:
      - appeal
  /appeal/{id}:
    delete:
      description: Удаление обращения
      parameters:
      - default: 1
        description: appeal id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Удаление обращения
      tags:
      - appeal
    get:
      description: Получение обращения по айди
      parameters:
      - description: appeal id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Appeal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Получение обращения по айди
      tags:
      - appeal
    put:
      description: Обновление обращения
      parameters:
      - description: appeal
        in: body
        name: appeal
        required: true
        schema:
          $ref: '#/definitions/input.AppealUpdate'
      - default: 1
        description: appeal id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Appeal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Обновление обращения
      tags:
      - appeal
  /appeal/{id}/comment:
    get:
      description: Получение комментариев обращения
      parameters:
      - default: 1
        description: appeal id
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: page
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: page
        in: query
        maximum: 20
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AppealCommentPaged'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Получение комментариев обращения
      tags:
      - appealComment
    post:
      consumes:
      - multipart/form-data
      description: Создание комментария
      parameters:
      - default: 1
        description: appeal id
        in: path
        name: id
        required: true
        type: integer
      - example: Текст кооментария
        in: formData
        name: text
        type: string
      - collectionFormat: csv
        description: photos
        in: formData
        items:
          type: file
        name: photos
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.AppealComment'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Создание комментария
      tags:
      - appealComment
  /appeal/{id}/status/{status_id}:
    post:
      description: Обновление статуса обращения
      parameters:
      - default: 1
        description: appeal id
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: new status id
        in: path
        name: status_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Обновление статуса обращения
      tags:
      - appeal
  /appeal_category:
    get:
      description: Получение всех категорий обращений
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.AppealCategory'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Получение всех категорий обращений
      tags:
      - appealCategory
  /appeal_category/{id}:
    get:
      description: Получение категории обращения по айди
      parameters:
      - description: appealCategory id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AppealCategory'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Получение категории обращения по айди
      tags:
      - appealCategory
  /appeal_category/{id}/appeal_types:
    get:
      description: Получение всех типов обращений по айди категории
      parameters:
      - description: appealCategory id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.AppealTypeByCategory'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Получение всех типов обращений по айди категории
      tags:
      - appealCategory
...
swagger: "2.0"